pyproject.toml
[tool.poetry]
name = "recipiesharbor"
version = "0.1.0"
description = ""
authors = ["logos201 <logos201@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aniso8601 = "9.0.1"
blinker = "1.6.3"
click = "8.1.7"
flask = "2.3.3"
flask-login = "0.6.2"
flask-migrate = "4.0.5"
flask-restful = "0.3.10"
flask-wtf = "1.2.1"
greenlet = "3.0.0"
gunicorn = "21.2.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
mako = "1.2.4"
markupsafe = "2.1.3"
packaging = "23.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
six = "1.16.0"
sqlalchemy = "2.0.22"
typing-extensions = "4.8.0"
werkzeug = "2.3.7"
wtforms = "3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
